Add dual-region configuration with site-to-site VPN
To showcase a more advanced Terraform configuration, introduce a
dual-region setup connected via site-to-site VPN.

To do so, define the two regions as "left" and "right" (to evoke
"local" and "remote", but without the ambiguity this creates based on
which region we're looking from). The "left" region is the one where
we already have resources; the "right" one is the new one we spin up
and connect to, via VPN.

To manage everything properly, define two provider aliases: the
"openstack" provider is unchanged from the prior configuration and
points to the Dx1 region. This corresponds to the "left" site. In
addition, define another alias for the openstack provider named
"right", using the Fra1 region.