Limit openstack provider to 1.53.x
The 1.54 version of the provider does work fine, however it produces
an utterly useless warning:

openstack_networking_network_v2.network_right: Destruction complete after 6s
╷
│ Warning: Argument is deprecated
│
│   with provider["registry.opentofu.org/terraform-provider-openstack/openstack"],
│   on providers.tf line 1, in provider "openstack":
│    1: provider "openstack" {
│
│ Users not using loadbalancer resources can ignore this message. Support for neutron-lbaas will be removed on next major release. Octavia will be the only supported method for loadbalancer
│ resources. Users using octavia will have to remove 'use_octavia' option from the provider configuration block. Users using neutron-lbaas will have to migrate/upgrade to octavia.
│
│ (and one more similar warning elsewhere)

There is nothing the learner/user can do about this. The warning
pretends that something is wrong with the configuration, when clearly
there isn't.

Thus, until this warning is fixed upstream, stick to the prior 1.53
version that does not have this problem.