fix: Set security group on port, not instance
Now that we pre-create the port before attaching it to an instance, we
have the option of referencing the security groups either from the
instance (as before), or from the port.

The former approach has the issue that it always results the instance
resource(s) always being "changed" upon apply, like so:

  # openstack_compute_instance_v2.instance will be updated in-place
  ~ resource "openstack_compute_instance_v2" "instance" {
        id                  = "f04e622e-9398-4d09-8040-b334452d6a60"
        name                = "instance-cc217_dev_250279fcee1ca1288b626f96f730a23a"
      ~ security_groups     = [
          - "secgroup-cc217_dev_250279fcee1ca1288b626f96f730a23a",
          + "ea489985-d94f-4c27-b947-8c753308a0c4",
        ]
        tags                = []
        # (15 unchanged attributes hidden)

        # (2 unchanged blocks hidden)
    }

Plan: 0 to add, 1 to change, 0 to destroy.

In other words, the security group (which the configuration already
references by ID) is somehow tracked internally by name, and is then
again "replaced" by the ID reference.

Setting the security group on the port avoids this issue.